requirements.txt
fastapi
tensorflow
face_recognition
numpy
pillow
deepface
python-multipart
uvicorn